complete chapter 2 question 2
