+++ Chap2Quiz.cpp
@@ -1,22 +1,10 @@
 // Quiz for chapter 2 of C++ tutorial on learncpp.com
 
 #include <iostream>
+#include "io.h"
 
 int readNumber();
 void writeAnswer();
-
-int readNumber()
-{
-    std::cout << "Enter an integer:";
-    int x{};
-    std::cin >> x;
-    return x;
-}
-
-void writeAnswer(int input)
-{
-    std::cout << "Your numbers added together are: " << input << ".\n";
-}
 
 int main()
 {
