+++ Chap2Quiz.vcxproj
@@ -149,6 +149,9 @@
   <ItemGroup>
     <ClCompile Include="Chap2Quiz.cpp" />
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="io.h" />
+  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
